Add partial matching for `dtw()` in python
SPRING [1][2]
is a streaming method for subsequence matching in data streams.
It differs from regular DTW in that it calculates partial matching.

In SPRING, the starting point for partial matching is obtained
by star-padding and STWM.
This commit adds `b_partial_start` flag, and if it is True,
`__dtw()` calculates the matching start point
by the star-padding-like and the STWM-like behavior.

In SPRING, the ending point for the partial matching is obtained
as the ending point of the optimal subsequence
as the data streaming progresses.
However, `__dtw()` does not handle data streams.
This commit adds `b_partial_end` flag, and if it is True,
`__dtw()` calculates the matching ending point as the point
with the lowest DTW distance.

[1]
Yasushi Sakurai, Christos Faloutsos, Masashi Yamamuro:
Stream Monitoring under the Time Warping Distance, ICDE2007,
https://doi.org/10.1109/ICDE.2007.368963

[2]
Yasushi Sakurai, Christos Faloutsos, Masashi Yamamuro:
Stream Processing under the Dynamic Time Warping Distance, DEWS2007,
https://www.ieice.org/iss/de/DEWS/DEWS2007/pdf/l6-5.pdf